pyproject.toml
[project]
name = "py-sandbox-11"
version = "0.1.0"
description = "Pure Async REST API with FastAPI, PostgreSQL, and clean architecture"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "httpx>=0.27.0",
    "pyjwt[crypto]>=2.10.0",
    "cryptography>=44.0.0",
    "passlib[bcrypt]>=1.7.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "testcontainers[postgres]>=4.9.0",
    "ruff>=0.14.4",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[project.scripts]
work_me = "mod_b.work_me:main"
api = "api.main:run"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
